Support Extra Bridges, VM Name and CPU Mode
Add 'cluster.nodes.worker.instances..extra_bridges' setting. Creates
additional network interfaces in the VM and maps them to a pre-configured
bridge on the host. A static IP address is added for each to the neplan
config. Child settings:
  bridge - name of the host bridge. E.g. br1
  network_interface - must match the name of the network link in the VM
    for this network interface. E.g. ens4
  ipCidr - IP address to assign in CIDR format. E.g. 192.168.0.100/24

Add 'cluster.nodes.worker.instances..name' setting. The VM name is formed
by appending this value to the cluster name. Default is to use the worker id.

Add 'cluster.nodeTemplate.cpuMode' setting. Sets the CPU mode used by the
VM (e.g. host-passthrough). Default: 'custom'.

Don't define the IP address when generating the libvirt_domain resource.
The IP is set by netplan config and this generates constant changes in the
terraform state.

Fix an error when generating the output if no IPs are set on a VM.